moved archive inside core and refactor
