--- ipol_demo/modules/archive/requirements.txt
@@ -1,7 +1,0 @@
-fastapi==0.92.0
-uvicorn==0.20.0
-pydantic==1.10.4
-configparser==3.5.0
-ipython==7.16.3
-python-magic==0.4.15
-requests==2.22.0
