+++ ipol_demo/modules/core/app.py
@@ -1,8 +1,10 @@
+from archive.router import archiveRouter
 from conversion.router import conversionRouter
-from core.router import coreRouter
 from demoinfo.router import demoinfoRouter
 from dispatcher.router import dispatcherRouter
 from fastapi import FastAPI
+
+from core.router import coreRouter
 
 app = FastAPI(root_path="/api")
 
@@ -11,3 +13,4 @@ app.include_router(coreRouter)
 app.include_router(demoinfoRouter)
 app.include_router(dispatcherRouter)
 app.include_router(conversionRouter)
+app.include_router(archiveRouter)
